Revert "Ensured that the weight in restart is float64"
This reverts commit 2de4cc006ef194d20ab46b0d79a0fb863d0527dc.

+++ fbpic/openpmd_diag/checkpoint_restart.py
@@ -245,11 +245,9 @@ def load_species( species, name, ts, iteration ):
     # Get the particles' momenta
     species.ux, species.uy, species.uz = ts.get_particle(
         ['ux', 'uy', 'uz' ], iteration=iteration, species=name )
-    # Get the weight (multiply it by the charge to conform with FBPIC,
-    # and make sure that the weight is obtained in float64, as this 
-    # was causing problems with older versions of openPMD-viewer)
+    # Get the weight (multiply it by the charge to conform with FBPIC)
     w, = ts.get_particle( ['w'], iteration=iteration, species=name )
-    species.w = species.q * w.astype(np.float64)
+    species.w = species.q * w
     # Get the inverse gamma
     species.inv_gamma = 1./np.sqrt(
         1 + species.ux**2 + species.uy**2 + species.uz**2 )
